Support XML and JSON inflector acronyms
Related to https://github.com/rails/activeresource/pull/361

Add special-case treatment of JSON and XML acronyms by declaring aliases
for the `JsonFormat` and `XmlFormat` constants.

In addition to the aliases, this commit also includes test coverage to
cover existing behavior.

The original PR included a comment citing that
`ActiveSupport::Inflector::Inflections#clear` was not working as
expected. It was resolved by [a76344f][], which was merged into `main`
prior to the `7.0.0` release. Since the CI matrix currently includes
`7-0-stable` as the minimum version, the code can rely on the resolved
behavior.

[a76344f]: https://github.com/rails/rails/commit/a76344ffc5a308e0ad6105fde921990f57fa0ea9