Base: Replace regexps in `method_missing` and `respond_to_missing`
Prior to this commit, the regular expressions used in both
`method_missing` and `respond_to_missing` were similar, but had slight
variations. That resulted in unknown attributes being able to be written
to through methods "did not respond to" (like `send("#{attr_name}=",
value)`). Similarly, the `?`-suffixed predicate methods would return the
same way when invoked.

Without the implementation changes, the following tests introduced in
this commit fail:

```
  1) Failure:
BaseTest#test_respond_to_unknown_attribute_writer [test/cases/base_test.rb:967]:
Expected #<Post:0x0000000120cb7958 @attributes={}, @prefix_options={}, @persisted=false> (Post) to respond to #unknown_attribute=.
```

This commit's changes replace the regular expressions with explicit
[String#end_with?][]. Originally (in [4179e98][] on May 4, 2007!), the
implementation was based on String comparison, but was eventually
replaced with a Regular-expression based implementation in [c7629a6][]
(on on Nov 9, 2009). The commit message for `c7629a6` cited <q>reliance
on string access core extension</q>. In the intervening years, Ruby's
standard library String implementation added `end_with?`, so the core
extensions mentioned in the message are no longer necessary in modern
Ruby versions.

Considerations
---

This change introduces subtle behavior changes. Whether or not it fixes
a bug or introduces a breaking change might be open to interpretation.

Conventions recommend that `respond_to?` and `method_missing` *should*
remain synchronized in what they consider inside and outside of a class
interface. Supporting the method invocations for `=` and `?` suffixed
methods without providing indications of that support when asked by
callers deviates from that convention.

In order to respond to `=`-suffixed writers and `?`-suffixed predicates,
the prior implementation required that the attributes being written to
or queried must have already been assigned to the internal `@attributes`
hash. That means that a `method_missing`-triggered write must've
pre-dated the `respond_to?` invocation.

Whether or not this is coincidental or intentional is a question best
asked of the historical authors and creators of the gem.

Ignoring intent, this commit proposes that both `respond_to?` and
`method_missing` communicate support for `=`- and `?`-suffixed methods
regardless of whether or not they've been written already.

[String#end_with?]: https://ruby-doc.org/3.4.1/String.html#method-i-end_with-3F
[4179e98]: https://github.com/rails/activeresource/commit/4179e98200b76c0e1d3a4b4ff0640bb363b3600c
[c7629a6]: https://github.com/rails/activeresource/commit/c7629a62d71a6e7c59b97a7b69066122664f9457